+++ CHANGELOG.md
@@ -1,3 +1,67 @@
+## 0.8.5
+
+* Bound
+
+## 0.8.4
+
+* Bound
+
+## 0.8.3
+
+* OTA
+
+## 0.8.2
+
+* Do not Disturb
+
+## 0.8.1
+
+* Reconnect
+
+## 0.8.0
+
+* Location
+
+## 0.7.9
+
+* stock market
+
+## 0.7.8
+
+* digiTop, digileft
+
+## 0.7.7
+
+* dial pointer
+
+## 0.7.6
+
+* weather, ota
+
+## 0.7.5
+
+* Pointer second
+
+## 0.7.4
+
+* Pointer second
+
+## 0.7.3
+
+* Alarm
+
+## 0.7.2
+
+* Alarm
+
+## 0.7.1
+
+* Alarm
+
+## 0.7.0
+
+* music
+
 ## 0.6.9
 
 * Sedentary